WORK ACCOUNT FLIP FLOPPING THE SCOPES!
Ok so this is a bit special. We flag endpoints as "work only lol". If they are work only then exclude them from scopes.
We include teams tools. If user tries to say list teams they should get auth error, where we pick this up
and add all WORK scopes. They then have to auth these, and if that works, hurray. That's the plan at least.